requirements.txt
Django~=3.2.10
djangorestframework~=3.11.1
celery[sqs]
django-celery-results
